Throwing water on grill makes fire stop spitting out of it
