+++ Assets/Prefabs/Cup.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d404c57a86f234445b24d449d9db902e
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
