+++ Assets/Scripts/BasePressurePlate.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 079292805827edc49b73eab833cd57f3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
