+++ Assets/Scripts/ItemPressurePlate.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 192a83e06d67fcf419ba7eab07aa3c8b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
